+++ src/AppView.tsx
@@ -20,7 +20,8 @@ import { SnackbarProvider } from 'notistack';
 // @ts-ignore:next-line
 import { registerSW } from 'virtual:pwa-register';
 
-import { App, MainPanes, PortState } from './App';
+import { App, MainPanes } from './App';
+import { PortState } from './Settings/PortConfiguration/PortConfiguration';
 import './App.css';
 import SettingsDialog from './Settings/SettingsView';
 import TerminalView from './Terminal/TerminalView';
@@ -65,11 +66,21 @@ const darkTheme = createTheme({
 });
 
 /**
- * Maps a port state to a colour used in the port state status background on the bottom toolbar.
+ * Maps a port state to the UI elements that should be used to represent it.
  */
-const portStateToBackgroundColor: { [key in PortState]: string; } = {
-  [PortState.CLOSED]: 'red',
-  [PortState.OPENED]: 'green',
+const portStateToToolbarStatusProperties: { [key in PortState]: any; } = {
+  [PortState.CLOSED]: {
+    color: 'red',
+    text: 'Port CLOSED',
+  },
+  [PortState.CLOSED_BUT_WILL_REOPEN]: {
+    color: 'orange',
+    text: 'Port CLOSED (will reopen)',
+  },
+  [PortState.OPENED]: {
+    color: 'green',
+    text: 'Port OPENED',
+  }
 };
 
 interface Props {
@@ -84,10 +95,12 @@ declare global {
 
 window.app = app;
 
+console.log('TEST')
+
 const AppView = observer((props: Props) => {
-  // const { app } = props;
 
   useEffect(() => {
+    console.log('useEffect1() called.');
     // We need to register the service worker AFTER the app
     // has rendered, because it we do it before we won't
     // be able to enqueue a snackbar to tell the user there
@@ -104,6 +117,13 @@ const AppView = observer((props: Props) => {
         app.swOnRegisterError(error);
       }
     })
+
+    const initFn = async () => {
+      console.log('Calling init()...');
+      await app.onAppUiLoaded();
+    }
+
+    initFn().catch(console.error);
 
     // Uncomment this if you want to test out the snackbar
     // for development reasons
@@ -281,7 +301,7 @@ const AppView = observer((props: Props) => {
             {/* Show port configuration in short hand, e.g. "115200 8n1" */}
             <Box>{app.settings.shortSerialConfigName}</Box>
             {/* PORT STATE */}
-            <Box sx={{ backgroundColor: portStateToBackgroundColor[app.portState], padding: '0 10px' }}>Port {PortState[app.portState]}</Box>
+            <Box sx={{ backgroundColor: portStateToToolbarStatusProperties[app.portState].color, padding: '0 10px' }}>{portStateToToolbarStatusProperties[app.portState].text}</Box>
           </Box>
         </div>