+++ src/index.tsx
@@ -7,20 +7,20 @@ import {
 } from "react-router-dom";
 import ReactGA from "react-ga4";
 
-import { App } from './App';
 import AppView from './AppView';
 import HomepageView from './Homepage/HomepageView';
 
-// Google Analytics
-ReactGA.initialize("G-SDMMGN71FN");
+// Google Analytics. Only initialize in production, otherwise things like
+// Playwright tests can spam GA and skew data
+if (import.meta.env.PROD) {
+  ReactGA.initialize("G-SDMMGN71FN");
+} else {
+  console.log('Detected dev. environment, not initializing Google Analytics.');
+}
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
-
-
-
-
 
 // Create routes. Only 2 routes. The root is the
 // landing page which is static, and then
@@ -37,19 +37,10 @@ const router = createBrowserRouter([
 ]);
 
 root.render(
-  <React.StrictMode>
+  // WARNING: StrictMode causes double renders, which causes problems
+  // during development when trying to open previously used serial ports
+  // and also with loading other things from local storage
+  // <React.StrictMode>
     <RouterProvider router={router} />
-  </React.StrictMode>
+  // </React.StrictMode>
 );
-
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://cra.link/PWA
-// serviceWorkerRegistration.unregister();
-// serviceWorkerRegistration.register();
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();